Set required_ruby_version
This is a new cop that is part of rubocop-govuk 4 that requires you to
specify the minimum Ruby version a gem supports. This cop makes a lot of
sense for gem distribution and clarity over which Ruby version is supported
for a gem.

It does however come at a cost of conflicting with a common GOV.UK
practice: GOV.UK typically sets the .ruby-version of gems to be a more
modern one than the minimum version the gem supports. Doing this however
conflicts with rubocop which expects (unsurprisingly) that the
ruby-version is the minimum one.

To reflect this I've changed this repo to use Ruby 2.6.6.